requirements.txt
alembic==1.7.5
asgiref==3.4.1
black==21.9b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.1
colorama==0.4.4
coverage==6.1.1
fastapi==0.68.2
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-spanner==3.11.1
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.12.0
idna==3.3
linecache2==1.0.0
Mako==1.1.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
proto-plus==1.19.8
protobuf==3.19.1
psycopg==3.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.6
python-dotenv==0.19.0
regex==2021.10.23
requests==2.26.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.3.23
sqlalchemy-spanner==1.0.0
sqlparse==0.4.2
starlette==0.14.2
tomli==1.2.2
traceback2==1.4.0
typing-extensions==3.10.0.2
unittest2==1.1.0
urllib3==1.26.7
uvicorn==0.15.0
